+++ integration/tests/common-extensions/counter.spec.ts
@@ -2,7 +2,7 @@
 import { Injectable } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { NgxsDataPluginModule } from '@ngxs-labs/data';
-import { DataAction, payload, StateRepository } from '@ngxs-labs/data/decorators';
+import { DataAction, Payload, StateRepository } from '@ngxs-labs/data/decorators';
 import { NgxsDataFactory } from '@ngxs-labs/data/internals';
 import { NgxsImmutableDataRepository } from '@ngxs-labs/data/repositories';
 import { NGXS_DATA_EXCEPTIONS } from '@ngxs-labs/data/tokens';
@@ -178,7 +178,7 @@ describe('[TEST]: CountState', () => {
             }
 
             @DataAction()
-            public withAction(@payload('val') val: number): void {
+            public withAction(@Payload('val') val: number): void {
                 this.ctx.setState(val);
             }
         }